Adding feature solving following issue:
Still, users often had trouble getting to mobile sites, even when companies offered them. The best approach is to auto-sense users' devices and auto-forward mobile users to the mobile site (even if they're using a high-end phone). You should also offer clear links from the desktop site to the mobile site, as well as a link back to the full site. As for link labels, we recommend "Mobile Site" and "Full Site," respectively.